+++ library/android-mvc/src/main/java/com/shipdream/lib/android/mvc/view/MvcFragment.java
@@ -18,6 +18,7 @@ package com.shipdream.lib.android.mvc.view;
 
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
+import android.support.v4.app.FragmentTransaction;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -86,6 +87,8 @@ public abstract class MvcFragment extends Fragment {
         /**
          * @return Indicates whether the fragment view is created when the fragment is created for
          * the first time. When this flag is true it's a good time to initialize the state fragment.
+         *
+         * <p>false will be returned when the view is created by rotation, back navigation or restoration</p>
          */
         public boolean isFirstTime() {
             return this.isFirstTime;
@@ -282,15 +285,16 @@ public abstract class MvcFragment extends Fragment {
             reason.isRotated = true;
         }
 
+        if (restoring) {
+            reason.isRestored = true;
+        } else if (!orientationChanged && !aboutToPopOut) {
+            //When the view is created not by orientation change nor poping out from back stack
+            reason.isFirstTime = true;
+        }
+
         if (aboutToPopOut) {
             reason.isPoppedOut = true;
             aboutToPopOut = false;
-        }
-
-        if (restoring) {
-            reason.isRestored = true;
-        } else if (!orientationChanged) {
-            reason.isFirstTime = true;
         }
 
         onViewReady(view, savedInstanceState, reason);
@@ -355,6 +359,16 @@ public abstract class MvcFragment extends Fragment {
      * back stack.
      */
     protected void onPushingToBackStack() {
+    }
+
+    /**
+     * Called when this fragment is being replaced by the nextFragment and right before the
+     * fragment transaction is committed. If transaction between fragments need to specify shared
+     * elements, do it here.
+     * @param transaction The transaction being committing
+     * @param nextFragment Next fragment is going to
+     */
+    protected void onPreTransactionCommit(FragmentTransaction transaction, MvcFragment nextFragment) {
     }
 
     boolean aboutToPopOut = false;